Handle the mapped but not added case
Sometimes on startup a window is mapped, but not yet added to the list of
windows that meta knows about. This means that it won't appear in the list
returned by list_windows(). However it also wouldn't get a mapped event,
because it was already mapped. Handle this window in _windowAdded and store
an array of known windows so that we don't handle other new windows in both
_windowAdded and _mapWindow

[endlessm/eos-shell#23]