Add a mechanism to directly open the apps page
Use this when the number of open windows drops to zero, but use the original workspaces page (with window picker) when the hot corner is activated.
This is a more correct version of commit 5c35502bbad870ef9eab2bb393b0095218e5d5bb, which reuses the existing infrastructure currently handling the Super+A shortcut (which has the same effect).