Use shell_window_tracker_is_window_interesting in WorkspaceMonitor
We use this already everywhere else, and we recently have had bugs where
windows were not being picked up correctly because we weren't using the
correct type check.
To minimize our delta and ensure consistency in behavior, change
WorkspaceMonitor to use the same function instead of its current
homebrew version of it.

[endlessm/eos-shell#176]