Show app page when toggling overview with no visible windows
Add a hideOrShowApps() method to Overview, that checks for the number of
visible windows before deciding whether to jump to the apps page or
actually hide the overview; the method is now called from both
Overview.toggle() and WorkspaceMonitor.updateOverview().

Since in many cases the shell code explicitly calls hide() on the
overview knowing a window will launch, the original hide() method is
left unchanged, so this only affects hot corner and keybindings.

[endlessm/eos-shell#176]