Resync button hover state before discarding event
When we click on another icon on the taskbar, the menu will be
dismissed, but since we hold a grab, the other icon won't get the enter
event to flip its hover state, and the clicked signal will not be
emitted as a consequence.
Use st_widget_sync_hover() to ensure the hover state is consistent
before passing the event down to the button again.

[endlessm/eos-shell#157]