Use the panel miniature as a target for the minimize animation
Mutter and the Shell support an icon-geometry property for a window,
which maps to the rectangle of its miniature on a panel. The
animation code for minimize will read that value and move the window
towards that point if it's set, so just set the icon geometry to the
allocation of the minuature whenever that changes.

The code path where icon-geometry is not set is also changed to minimze
to (0, monitorHeight) instead of (0, 0) for consistency.

[endlessm/eos-shell#185]