Add support for a hint-actor to StEntry
Previously, StEntry supported setting a text string hint, through the
hint-text property.
This commits adds a similar support for a generic actor to be used as a
hint for the text entry, following the same logic as the text hint. In
case both the actor and the text are set, the actor will take
precedence.

[endlessm/eos-shell#86]