Allow DnD from popup folders
If an icon is dropped on the Trash, a dialog is opened to ask if icon must
be removed.

If an icon is dropped in the main view, it is removed from the folder
and inserted in the dropped place in the main view.

If an icon is dropped over a Folder icon, it is removed from the current
folder and inserted in the new folder.

If an icon is dropped in the Dash view, it is added as favourite.

At the moment, dragging/hovering an icon in a different place inside the
folder does nothing.

[endlessm/eos-shell#256]