Fix cropped folder views at low resolutions
Streamline this code much closer to the current upstream gnome-shell
one. This allows us to avoid ClutterConstraints to position the folder
popup, which makes the whole flow much more understandable.
Now resizing of the view is always forced by setting the 'y' property on
the folder popup, which works correctly at all resolutions.
[endlessm/eos-shell#240]