Do not use overview.toggle() with the overlay-key
When toggling via keyboard we want to check that we're not in the Apps
page of the overview; in order to do this, we use a separate method on
the overview, instead of complicating the logic inside toggle(). The
toggle() method is getting called from multiple sites, and hunting them
all down is bound to trigger a cascade of regressions.

[endlessm/eos-shell#177]