Don't trigger a whole redisplay when setting the same text
When the text is edited with the same string as the old one, there's no
need to actually go and rename the desktop/directory file.
Treat that case as a cancelled edit instead.

This commit also adds some code that makes sure the hover state of the
app icon button is consistent when we exit the grab in place while
editing.

[endlessm/eos-shell#448]