Fix click action on the text surroundings
Previously, when clicking on the surroundings of the icon text, but
still inside the text box, we would implicitly activate the icon action.
Since the text box actor is grabbing, always listen for button press
events on it, and eat events to avoid the default action from being
activated.
This also requires refactoring the current label state in a separate
variable, which cleans up the code a bit.

[endlessm/eos-shell#448]