Simplify internal state handling of ViewSelector
Instead of changing the Show Apps button state whenever we're entering
or exiting the overview, change code so that the state of the button is
set from _showPage(), and not viceversa.
This way, when show() calls _showPage(this._workspacesPage), the button
will automatically be reset without any extra overhead, as _showPage()
already shortcircuits same-page changes, and we can greatly simplify the
code handling the internal state related to that.

[endlessm/eos-shell#268]