Fix a warning when closing a folder popup
We used to cache the popup actor after it was dismissed, so it was
necessary to tie its lifecycle to the source icon - since right now we
always destroy it after it closes, doing this might cause destroy() to
be called on an invalid object, triggering a console warning.

Just remove the link between the source icon and the popup actor to fix
the warning.

[endlessm/eos-shell#468]