Don't relayout the whole view when moving icons in a folder
In order to achieve this, we now move the code relayouting the view to
be entirely in the base class. Code populating the view will then run
only for that one folder, that was reordered, leaving the popup in
place.
In addition, this allows us to unify all the code paths that add items
to the view under a single addIcons() method on the base class.

[endlessm/eos-shell#468]