Store last cursor position instead of the onIcon state
this._onIcon is actually duplicated in (this._onIconIdx != -1). It's
more interesting to find out when the cursor location changes to reset
nudge transforms. This way we can catch the case when, inside the same
view, we move an item towards an empty area and then back towards the
grid. When that happens, we will get a different cursor location, and we
will nudge accordingly.

[endlessm/eos-shell#99]