Applications minimize on taskbar icon click
The currently focused application minimizes when its icon is clicked. However,
if the app has multiple windows a popupmenu appears to display them, thus
avoiding minimization (similar to Windows).

Also, the minimization animation currently pulls the window to the lower left
corner of the screen, whereas the activation animation causes the window to
fade in. Presumably these animations should match.