Use a custom layout manager for AppDisplay
The layout manager changes the allocation for the container of the
application grid and the entry so that the latter is always allocated as
if there were four rows of icons in the grid, distributing the extra
margin along the vertical sides of the entry as per-designs.
In case that's not possible, for instance because the view has more
rows, the entry will still reserve a minimum padding around it, which
can be controlled from CSS using a custom style property.

[endlessm/eos-shell#62]