Add deletion of user-generated .desktop files
User-generated .desktop files are generated when the user renames
an icon on the desktop.  The .desktop files are deletable if they are
in the user's home folder ($XDG_DATA_HOME).

With this change, deletion of the .directory files uses the same logic.
So, it is no longer necessary to prefix the user folders with
'userdir-'.

[endlessm/eos-shell#432]