Consider favorite placeholder when computing "y"
When the placeholder is present, the "y" already depends on the
favorites icons and the placeholder. But when it's null, it only depends
on the favorites icons, though as we are adding the placeholder, we need
to take that in account: if placeholder were present, it would be
increasing the "y" position by the size of the placeholder. So let's do
it manually.

[endlessm/eos-shell#347]