Minimize other windows upon close/minimize
If the current window is maximized, closing or minimizing it
minimizes all other windows to display the app selector.
If the current window is not maximized, leave the other windows alone.

A user that works with all default maximized windows will get behavior
similar to a smartphone, where the minimize button essentially
always takes the user home to the desktop app selector.

A power user that wants to display more than one window at a time
can do so by resizing after the default open maximized.
In that case closing a resized window will not impact
the other windows that are currently visible.

The effect is the same, regardless of how the app is opened
(from the app selector, window picker, taskbar, or file manager).

[endlessm/eos-shell#587]