Use custom icon for missing app icon
The required icon (generic-app.png) is provided by the eos-theme package.
The designers also specified a generic-link.png, but at the moment there
is no mechanism in place to know whether the missing icon was for an app
or for a link, so we use the more generic missing app icon.

It would not be appropriate to re-theme the application-x-executable icon,
since that icon may have other meanings than just a missing icon.

[endlessm/eos-shell#607]