Show view pages directly when entering overview
Instead of havign a deferred mechanism in place, use a targetPage
property in Overview. Together with a change to ViewSelector.show()
taking the page directly, this ensures we get propagate the right state
to ViewSelector without any roundtrips between that and Overview.

The additional deferred mechanism is not needed because
Overview._animateVisible() will always be re-entered when a show request
is received during the hide animation, and at that point we will ensure
the correct target page throught the new property.

[endlessm/eos-shell#421]