Use custom icon for missing link icon
The required icon (generic-link.app) is provided by the eos-theme
package.

Links are treated as applications: they are launching using epiphany
browser.

So in order to know if the app is actually an app or a link, we check
the command-line: if it is epiphany, then it is a link, and we return
generic-link.png; otherwise, it is an application and we return
generic-app.png.

Note that above check is done after we know the app has not a proper
icon, so we need to return the default one.

[endlessm/eos-shell#608]