Make the animation for the greeter consistent with panel position
As the position of our panel is at the bottom, we need to translate it
from below, and not above.
Also, we need to hide the trayBox below it for the duration of the
transition, as the regular session does. The trayBox will be shown again
in the completion callback for the startup animation.

[endlessm/eos-shell#632]