Create invisible pages by default
Commit dd22271ec3af02d6390544e4102e02909376a2fa changed page creation in
ViewSelector to be transparent by default. That works to hide the
workspaces page when entering the overview directly to the app
selector, but the actor will still be mapped.
Since we have a ClickAction bound to the mapped actor state of the
workspaces page to catch clicks on empty space, the action will be
incorrectly enabled initially when entering the overview directly to the
app selector.

This commit fixes the bug by also making pages invisible by default on
creation.

[endlessm/eos-shell#421]