Start message tray below the panel
Return the offscreen position of the message tray to be
completely below the bottom panel, and then animate it
all the way through the panel when activated.

Since the bottom panel is now semi-transparent,
we can't hide the message tray beneath the panel
without impacting the display of the panel.

[endlessm/eos-shell#685]