Disable background menu
The background menu was not being displayed for the app selector
or the window picker, so at least for now, we disable it altogether
for consistent behavior.

This works around an issue where the background menu could be
activated by a right-click on the shell's full-screen splash page,
if the application has its own splash screen in the background.

If this feature is to be renabled in the future, it should be
available on the app selector and window picker, but not while
the full-screen splash page is displayed.

[endlessm/eos-shell#720]