Move status indicators closer together
Made the layout of status indicators on the right of the panel
roughly match that of the mini icons on the left of the panel.
In order to keep the center-to-center distance roughly the same,
a spacing of 12px between status indicators seemed to match a
spacing of 10px between mini icons (for 24x24 mini icons).

Though we don't have exact specs for this spacing,
this looks much closer to the design flats than before.

[endlessm/eos-shell#755]