Clean up CSS for search hover/focus
Although they currently have the same contents,
so there is no noticeable impact for this change,
if the hover and focus state had different contents
the order matters.  The focus needs to be specified
after the hover to avoid odd displays of hover vs. focus
when hovering over/off the focused widget.

Also, removed the transition duration of 0ms for focus.
The transition duration is ignored here, so let's not
specify a value that may confuse someone into thinking
it could be adjustable.

[endlessm/eos-shell#770]