Make sure to destroy the original icon when destroyed
When in a DnD operation, we substitute our our icon with the
custom "drag begin icon". As dnd.js relies on the initial actor being
destroyed in order to automatically cancel the drag, and the initial
actor is saved in this._origIcon, we have to make sure we call destroy()
on it at the same time the actor is destroyed, as this._origIcon won't
be in the hierarchy being destroyed otherwise.

[endlessm/eos-shell#780]