Set a flag when a custom name is added for an icon
ShellAppSystem refreshes its internal state based on file system
notifications when any of the desktop files are added/removed/changed
from one of the known locations.

As rebuilding the full state after a FS notification will take a while,
it's possible to quickly rename an icon again before the notification
for a previous rename comes by.
In such a case we will still get another notification for the third name
afterwards, but since we're destroying the view state when the name
doesn't match, and the ShellApp one is preferred, we will show the
outdated name in the meantime.

This commit fixes this bug by setting a flag on an icon after it's
renamed. The flag will be cleared when the new name coming from the
ShellApp state matches the custom one, but not otherwise.

This still leaves open a theoretical race condition if the .desktop file
is modified out of the shell while we're waiting for a notification
immediately after a rename, but the case should be rare enough not to
worry about.

[endlessm/eos-shell#780]