Make sure our destroy callback is executed before GrabHelper's
grabHelper.addActor() will add a callback to the destroy signal of the
label actor, which will remove it from the grab stack.
We want our callback to be executed before it, so we have a chance to
cancel editing and ungrab the actor, keeping the grab state consistent.

This commit fixes the bug by connecting to 'destroy' instead of
overriding its vfunc, and by moving the signal connection before calling
grabHelper.addActor() in EditableLabel.

[endlessm/eos-shell#780]