Fix glitch in popup submenu open/close
Make sure that the change to the scrollbar policy during open
is reverted during close.

The existing code adjusted the scrollbar policy upon opening,
after showing the actor.  This caused problems where alternate
opens of the popup submenu would kill the menu due to showing
the actor with the policy set from the previous open.
I think this problem may only occur if using CSS
to specify the max-height of the popup submenu.

[endlessm/eos-shell#681]