Revert "popupMenu: Don't always ignore SubMenu children in width requests"
This reverts commit 2ed28211ed8904e1f926020e5273bd2494770315.

This upstream change does not work properly when the submenu children
have any CSS, as the CSS is not applied before the submenu is opened.
This was causing glitches in the display of the network status,
where the incorrectly calculated submenu column widths would be applied
to the displayed status.  Then, once the submenu is opened, it would be
resized to the correct width, but sometimes only after moving the cursor
after the submenu has been opened.

In talking with Cosimo, we agreed that the original behavior that the
upstream commit was trying to avoid is better than the incorrect sizing
that resulted from the change.  It appears that unless clutter were to
take into account the CSS when calculating the preferred width of
submenus that are not yet open, this is the best we can do for now.

[endlessm/eos-shell#681]