StButton: sync hover state before checking it
As documented in st_widget_set_track_hover(),
st_widget_get_hover() may not report accurate results when
another actor has a grab. Sync the hover state before
checking it to avoid issues.

This fixes an issue with the Endless Mobile ViewIcon class:
Clicking on the icon label to edit it causes a pointer grab to happen.
Then when moving the mouse to a StButton, the StButton is not aware of
the pointer movement. Now pressing the mouse button did not register
a click, since the StButton did not believe the mouse was hovered over
it.

[endlessm/eos-shell#940]

[endlessm/eos-shell#940]