Add a clone of the application grid to the desktop
We clone the applications page from ViewSelector, add a top panel ghost
similar to how Overview would do - to get its sizing right - and add it
on top of the desktop background actor for the primary monitor.
Upon click on any space in the background, we seamlessly transition to
the overview applications page.
To make this transition smooth, as the clone is desaturated and
semi-opaque, we need to also add a saturation effect to the overview
actor - which we only enable while the transition is in progress.