Show app store even if overview already hidden
In the case where the overview is already hidden,
we need to directly show the app store page
rather than try to hide the overview and connect
to its hidden signal.

The d-bus API was updated such that ShowPage takes care
of showing the app store, so that the shell no longer
needs to separately select the page and then toggle it visible.

[endlessm/eos-shell#932]