User menu stays open when hovering over separator or app mini icons
Hovering over any panel entry associated with a PopupDummyMenu, i.e.
an entry without a menu that is directly part of the panel (which is
just the AppIconBar at the moment) doesn't cause the perviously open
menu to close. The mini icon labels that appear on hover are also
raised to the top so they appear on top of the user menu. Clicks on
app mini icons, unlike hover events, cause the user menu to close.

[endlessm/eos-shell#992]