Configure hot corner location via gsettings
Previously, the position was based on the text direction.
Left-to-right text would have the hot corner in the upper-left corner.
Right-to-left text would have the hot corner in the upper-right corner.

Instead of using the text direction, we now use a gsetting.
The default location is in the upper-left corner.

[endlessm/eos-shell#944]