Unset the key focus we set when hovered
When a button gets hovered, we also set the key focus on it, to make it
activatable by keyboard. If we are unhovering a button, and it had the
key focus, we also should move the key focus out of it, or it will keep
appearing focused after we moved out of it.

[endlessm/eos-shell#1019]