Fix premature skype exits
Previously, we simply killed Skype when the user closes a window.
This inadvertantly killed the application when the user first logs in,
since entering login credentials closes the window.
Also, the close of popups like the phone dialer or notifications
would kill the application.

Instead, when one window is closed, we wait some amount of time
and then check if there are any other Skype windows open.
If not, then we kill the application.

During testing on a development machine, a timeout of 80 ms
was sufficient.  Let's use a timeout of 1000 ms to provide some margin.

[endlessm/eos-shell#938]