Combine power off and log out to single button
Replaces two buttons with a single button labeled "EXIT".
This new button is merely a new label for the power off button.
The power off confirmation dialog adds an extra button
to provide the option to log out.

Implement via a gsetting, so the original behavior
can easily be restored with the following:
gsettings set org.gnome.shell separate-power-off-log-out true

[endlessm/eos-shell#1010]