Fix app store activation from empty desktop
Currently, we rely on the remote visibility notification in order to
know whether we need to call toggle().
This causes a race condition if showApps() is called immediately after a
call to toggle(), since the app store is scheduled to be toggled again
when visible and the overview is showing. The end result is the app
store will be visible when we expect not to, causing it to hide instead
of show next time.
This commit fixes this by caching the visible state in the app store
object instead.

[endlessm/eos-shell#1169]