Only maximize windows for spawned applications
Previously, we would maximize any window created on the MetaDisplay.
That caused us to maximize windows that we shouldn't have in the first
place, like preferences windows or property dialogs.
The idea here is to ensure we only maximize primary application windows
instead. To achieve that, we listen to a signal GIO will emit when
lauching an application, and keep track of the launched ShellApp.
Only in case the next window we get matches the last launched
application we proceed to maximizing it.

[endlessm/eos-shell#1210]