Create user local applications/folder directories at startup
GMenu won't install a file monitor for those dierctories if they don't
exist, which is the default after the first login (until the user plaus
with the app store actually).
That causes applications/links added from the app store after first
login to be ignored until the next Shell reboot, as by that moment, file
monitors will be installed for the directories that now exist.

This commit ensures those directories exist when the Shell starts up, so
they will always be monitored by GMenu.

[endlessm/eos-shell#1382]