Work around for the minimize hang
This is hopefully a temporary fix for the occasional hang in the
shell when minimizing an app. When we discover the actual cause and
fix it, this should be reverted.

The temp fix is to only call overview.showApps() after window minimize
is complete. So rather that animating the icon grid active at the
same time as animating a window minimize, these will happen sequentially
[endlessm/eos-shell#1481]