Make gnome-shell-extension-prefs a binary executable
Since commit 1ebb162a00bf6c4b0 moved JS sources into resources,
the extension-prefs tool was broken. To fix it, we would either
need to generate an external GResource in addition to the generated
C code and teach gjs-console about loading it before evaluating
the script, or turn gnome-shell-extension-prefs into a binary with
the JS resources compiled in.

https://bugzilla.gnome.org/show_bug.cgi?id=722334

Conflicts:
	src/Makefile.am
	src/eos-shell-extension-prefs.in