Avoid tracking idle time for password dialog reset
Currently, we track idle time and reset the lock screen after 2
minutes of inactivity. When that happens, we enter an infinite recursion
loop which will make the login entry flash on the screen, and eventually
eat all the memory in the system. Resetting the lock screen will
recreate another unlock dialog, which in turn will create another
MetaIdleMonitor watch, which will fire right away since we're already
idle, or we wouldn't have reset the lock screen.
This wasn't a problem before the 3.10 rebase because GnomeIdleMonitor
didn't fire an event right away when a watch was added with a timeout
already past the idle time.

This could be fixed by moving the MetaIdleMonitor watch to ScreenShield,
so to avoid recreating the watch every time, but it turns out the only
purpose of resetting the lock screen was to make the curtain re-appear.
As we don't have the curtain, fix the bug by removing the idle monitor
altogether.

[endlessm/eos-shell#1653]

[endlessm/eos-shell#1653]