Automake integration for Jasmine tests.
This commit integrates the Jasmine JS tests with the automake
parallel-tests driver.

The jasmine test driver will only open files that have the name
"test" in them. As such the tests needed to be renamed.

In order to run tests individually without automake involvement,
the runner.js script inside the Jasmine submodule can be invoked
directly. Note that the environment must be set up correclty
in order to use this, particular GI typelib paths.