Dispose cairo context in osdWindow
Need to manually dispose of cairo contexts used in gjs with $dispose(),
or the context object will leak. This class used cairo for drawing but
were missing the dispose call.

Upstream included fix for the screen guard, if we every update our
screen guard to new one that uses cairo we should cherry pick that
dispose call as well

https://bugzilla.gnome.org/show_bug.cgi?id=722812