Hide extra workspace actors when parent are hidden
We add extra workspaces directly to the LayoutManager's OverviewGroup,
so we manually propagate the view's opacity to the workspace actor, as
they're not in the same hierarchy.
At the same time though, we forgot to propagate the actor's visibility
too. The result is the dropRect of the extra workspace actor would stay
mapped even when the workspaces page was hidden, interfering with DND.

This commit syncs visibility between the view actor and the extra
workspace, fixing the erroneous DND picking.

[endlessm/eos-shell#1855]