Remove BuildDepends on gdm and associated postinst script
Seems to cause a circular dependency from eos-shell to gdm to
gnome-session back to eos-shell that OBS can't handle. It was added
back in da690294 to support creating links from gdm private files to
eos-shell's directory, but now gdm3 installs it's library and
introspection data in the system standard locations. The Depends on
gdm remains as in upstream to allow files in gdm owned directories to
be installed.

[endlessm/eos-shell#1851]