Fix undo deletion of folders
Main.overview.setMessage() will call both the destroy and the undo
callback when the undo action is triggered.
Since the canDelete() check will pass for custom folders (they're files
in the home directory). we will effectively delete the folder file right
after adding it back to the grid layout, which will of course make it
disappear immediately.
Fix this by tracking whether we were undone before deleting the item.
[endlessm/eos-shell#2046]

[endlessm/eos-shell#2046]