Consistently use this._app in _onAppStateChanged
If one application fails to launch, and we clear the splash screen
without ever having seen one window from it, we will not disconnect the
signal to app-state-changed until the next window shows up, since we
don't know if we ever will get a window.
That's fine, as long as the function only acts on local state of the
AppActivationContext - in this case instead we would have quit the
later-spawned app instead.

[endlessm/eos-shell#2012]