Handle LibreOffice launch as special case
Launching LibreOffice is special when it comes up with the recovery
page.

Due the recovery page is shared by all the libreoffice applications, and
thus it has a different wmclass, plus LibreOffice forks before launching
the windows, makes it difficult for the shell to correctly identify the
recovery window as belonging to the launched libreoffice application,
and thus it is reported as a different application.

We deal with this special deal case identifying the recovery application
as being the libreoffice one, and thus closing the splash screen
correctly.

[endlessm/eos-shell#2238]