Show new windows in the Windows Overview
When the Windows Overview is shown, and a new window comes up, it is
checked if the new window must be shown in the overview.

But a small typo in the check condition was avoiding it.

This commit fixes it, so as soon as a new window comes up it is shown
in the windows overview so user can select it.

[endlessm/eos-shell#1945]