StEntry and StLabel: re-create the shadow when opacity has changed.
The shadow is created by drawing the widget to an offscreen texture,
but if the widget's opacity is 0, the shadow will be transparent
until an (unrelated) style change invalidates it at a moment when
the widget happens to be opaque.

This was fixed by using CLUTTER_OFFSCREEN_REDIRECT_ALWAYS in widgets
that had a text shadow. However, this solution causes glitches when
we create a Clutter.Clone of the desktop and move it to reveal the
message tray.

[endlessm/eos-shell#1630]