Window Manager: use an overlay to redirect clicks to the desktop.
Calling windowActor.hide() is not enough to really hide a window:
it will still be there, invisible but able to receive clicks that
should be going to the desktop.

We work around this problem by using an invisible overlay that
covers the part of the work area outside of the SideComponent;
clicks on this overlay will cause the emission of the
'background-clicked' signal.

[endlessm/eos-shell#1003]