appActivation: remove Libreoffice special case
This goes against the logic explained in the comment a few lines above:
since we can't fully trust the app state changed signal to be emitted on
an application with the same WMClass as the one we launched, we pick the
first application that flips its state to running to determine when
closing the splash screen.

The special case for Libreoffice instead, was making it so that when we
got an application not matching what we had launched (and it wasn't
Libreoffice itself), we would leave the splash screen hanging.

This commit still preserves part of the fix for the Libreoffice recovery
screen, so that doesn't regress.

[endlessm/eos-shell#2476]