workspaceMonitor: track windows in destruction
We now wait for "destroy-completed" in order to show back the
overview; we'll get a "window-removed" signal between "destroy" and
"destroy-completed" though, which will remove the window from our known
set of windows.
We have then to track windows in destruction in a separate HashMap, and
make sure to only call back into showing the overview when we have
destroyed a known window.

[endlessm/eos-shell#2730]